requirements.txt
# Smart Absensi - Requirements
# Python 3.11+ required

# Core libraries
numpy==1.26.4
opencv-python>=4.11.0
pandas>=2.3.0
pillow>=11.3.0

# Face recognition
face-recognition>=1.3.0
dlib>=19.24.0
cmake>=3.25.0

# Web framework (FastAPI)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Security & Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-dateutil>=2.8.2

# Jupyter
jupyter>=1.0.0
notebook>=7.0.0

# Optional (for Excel export)
openpyxl>=3.1.0
